+++ src/lib/Visualization/Molecule/BallsAndSticks/Geometry.ts
@@ -5,6 +5,10 @@
 namespace LiteMol.Visualization.Molecule.BallsAndSticks {
     "use strict";
 
+    const enum Constants {
+        MetalDashSize = 0.15
+    }
+
     namespace BallsAndSticksHelper {
         export function analyze(model: Core.Structure.Molecule.Model, atomIndices: number[], params: Parameters) {            
             const bonds = Core.Structure.computeBonds(model, atomIndices, {
@@ -20,57 +24,64 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
                 lastResidue = raI;
             }
 
-            let stickCount = 0;
-            let { type, count } = bonds;
+            const metalDashFactor = 1 / (2 * Constants.MetalDashSize);
+
+            let covalentStickCount = 0, metallicStickCount = 0;
+            let { type, count, atomAIndex, atomBIndex } = bonds;
+            let { x, y, z } = model.positions;
             for (let i = 0; i < count; i++) {
                 const t = type[i];
-                if (t === 0 || t === 5) stickCount += 1;
-                else if (t >= 1 && t <= 4) stickCount += t;
+                if (t === 0) covalentStickCount += 1;
+                else if (t >= 1 && t <= 4) covalentStickCount += t;
+                else if (t === 5) {
+                    const a = atomAIndex[i], b = atomBIndex[i];
+                    const dx = x[a] - x[b], dy = y[a] - y[b], dz = z[a] - z[b];
+                    const len = Math.sqrt(dx*dx + dy*dy + dz*dz);
+                    metallicStickCount += Math.ceil(metalDashFactor * len);
+                }
             }
 
             return {
                 bonds,
-                stickCount
+                covalentStickCount,
+                metallicStickCount
             }
         }
     }
 
     import BT = Core.Structure.BondType
+    import Geom = Core.Geometry
+    import Vec3 = Geom.LinearAlgebra.Vector3
+    import Mat4 = Geom.LinearAlgebra.Matrix4
+    import GB = Utils.GeometryBuilder
 
     class BondModelState {
-        atomsVector = new THREE.Vector3();
-        center = new THREE.Vector3();
-        rotationAxis = new THREE.Vector3();
-        upVector = new THREE.Vector3(0, 1, 0);
+        rotationAxis = Vec3.zero();
+        bondUpVector = Vec3.fromValues(1, 0, 0);
+        dashUpVector = Vec3.fromValues(1, 0, 0);
+        dir = Vec3.zero();
 
-        scaleMatrix = new THREE.Matrix4();
-        rotationMatrix = new THREE.Matrix4();
-        translationMatrix = new THREE.Matrix4();
-        offsetMatrix = new THREE.Matrix4();
-        finalMatrix = new THREE.Matrix4();
-
-        sticksDone = 0;
+        scale = Vec3.zero();
+        translation = Vec3.zero();
+        rotation = Mat4.zero();
+        
         radius = 0.15;
-        offset = new THREE.Vector3();
-        a = new THREE.Vector3();
-        b = new THREE.Vector3();
+        offset =  Vec3.zero();
+        a = Vec3.zero();
+        b = Vec3.zero();
 
         constructor(
-            public template: THREE.BufferGeometry,
-            public templateVB: Float32Array,
-            public templateNB: Float32Array,
-            public templateIB: Uint32Array,
-            public templateVertexCount: number,
-
-            public vertices: Float32Array,
-            public normals: Float32Array,
-            public indices: Uint32Array) {
+            public bondTemplate: Geom.Surface,
+            public cubeTemplate: Geom.Surface,
+            public builder: GB) {
         }
 
     }
 
-    class BuildState {
-        static getBondTemplate(radius: number, tessalation: number) {
+    namespace Templates {
+        const bondCache: {[t: number]: Core.Geometry.Surface} = {};
+        export function getBond(tessalation: number) {
+            if (bondCache[tessalation]) return bondCache[tessalation];
             let detail: number;
 
             switch (tessalation) {
@@ -83,14 +94,16 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
                 default: detail = 14; break;
             }
 
-            let template = GeometryHelper.getIndexedBufferGeometry(
-                new THREE.LatheGeometry([new THREE.Vector3(0, radius, -1 / 2), new THREE.Vector3(0, radius, 1 / 2)], detail, Math.PI));
-            template.applyMatrix(new THREE.Matrix4().makeRotationAxis(new THREE.Vector3(1, 0, 0), -Math.PI / 2));
-            return template;
+            const geom = new THREE.TubeGeometry(new THREE.LineCurve3(new THREE.Vector3(0, 0, 0), new THREE.Vector3(1, 0, 0)) as any, 2, 1.0, detail);
+            const ret = GeometryHelper.toSurface(geom);
+            bondCache[tessalation] = ret;
+            return ret;
         }
 
-        static getAtomTemplate(radius: number, tessalation: number) {
-            let base: any;
+        const atomCache: {[t: number]: Core.Geometry.Surface} = {};
+        export function getAtom(tessalation: number) {
+            if (atomCache[tessalation]) return atomCache[tessalation];
+            let base: any, radius = 1;
             switch (tessalation) {
                 case 0: base = new THREE.OctahedronGeometry(radius, 0); break;
                 case 1: base = new THREE.OctahedronGeometry(radius, 1); break;
@@ -100,73 +113,52 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
                 case 5: base = new THREE.OctahedronGeometry(radius, 3); break;
                 default: base = new THREE.IcosahedronGeometry(radius, 3); break;
             }
-            return GeometryHelper.getIndexedBufferGeometry(base);
+            const ret = GeometryHelper.toSurface(base);
+            atomCache[tessalation] = ret;
+            return ret;
         }
 
+        let dash: Core.Geometry.Surface | undefined = void 0;
+        export function getDash() {
+            if (dash) return dash;
+            dash = GeometryHelper.toSurface(new THREE.BoxGeometry(1, 1, 1));
+            for (let i = 0; i < dash.vertices.length; i += 3) {
+                dash.vertices[i] += 0.5;
+            }
+            return dash;
+        }
+    }
+
+    class BuildState {
         tessalation = this.params.tessalation;
         atomRadius = this.params.atomRadius;
         bondRadius = this.params.bondRadius;
         hideBonds = this.params.hideBonds;
 
-        bondTemplate = BuildState.getBondTemplate(1.0, this.tessalation!);
-        atomTemplate = BuildState.getAtomTemplate(1.0, this.tessalation!);
-        cubeTemplate = GeometryHelper.getIndexedBufferGeometry(new THREE.BoxGeometry(1, 1, 1));
+        bondTemplate = Templates.getBond(this.tessalation!);
+        atomTemplate = Templates.getAtom(this.tessalation!);
+        dashTemplate = Templates.getDash();
 
-        bondTemplateVertexBuffer = (<any>this.bondTemplate.attributes).position.array;
-        bondTemplateVertexBufferLength = this.bondTemplateVertexBuffer.length;
-        bondTemplateVertexCount = (this.bondTemplateVertexBufferLength / 3) | 0;
-        bondTemplateIndexBuffer = (<any>this.bondTemplate.attributes).index.array;
-        bondTemplateIndexBufferLength = this.bondTemplateIndexBuffer.length;
-        bondTemplateNormalBuffer = (<any>this.bondTemplate.attributes).normal.array;
-
-        cubeTemplateVertexBuffer = (<any>this.cubeTemplate.attributes).position.array;
-        cubeTemplateVertexBufferLength = this.cubeTemplateVertexBuffer.length;
-        cubeTemplateVertexCount = (this.cubeTemplateVertexBufferLength / 3) | 0;
-        cubeTemplateIndexBuffer = (<any>this.cubeTemplate.attributes).index.array;
-        cubeTemplateIndexBufferLength = this.cubeTemplateIndexBuffer.length;
-        cubeTemplateNormalBuffer = (<any>this.cubeTemplate.attributes).normal.array;
-
-        atomTemplateVertexBuffer = (<any>this.atomTemplate.attributes).position.array;
-        atomTemplateVertexBufferLength = this.atomTemplateVertexBuffer.length;
-        atomTemplateVertexCount = (this.atomTemplateVertexBufferLength / 3) | 0;
-        atomTemplateIndexBuffer = (<any>this.atomTemplate.attributes).index.array;
-        atomTemplateIndexBufferLength = this.atomTemplateIndexBuffer.length;
-        atomTemplateNormalBuffer = (<any>this.atomTemplate.attributes).normal.array;
-
-        atomBufferSize = this.atomTemplateVertexBufferLength * this.atomIndices.length;
-        atomVertices = new Float32Array(this.atomBufferSize);
-        atomNormals = new Float32Array(this.atomBufferSize);
-        atomTriangleIndices = new Uint32Array(this.atomTemplateIndexBufferLength * this.atomIndices.length);
-        atomColors = new Float32Array(this.atomBufferSize);
-        atomPickColors = new Float32Array(this.atomTemplateVertexCount * 4 * this.atomIndices.length);
+        atomVertexCount = this.atomTemplate.vertexCount * this.atomIndices.length;
+        atomTriangleCount = this.atomTemplate.triangleCount * this.atomIndices.length;
+        atomBuilder = Utils.GeometryBuilder.create(this.atomVertexCount, this.atomTriangleCount);
+        atomColors = new Float32Array(this.atomVertexCount * 3);
+        atomPickColors = new Float32Array(this.atomVertexCount * 4);
 
         atoms = this.model.data.atoms;
         positions = this.model.positions;
         cX = this.positions.x; cY = this.positions.y; cZ = this.positions.z;
         atomSymbols = this.atoms.elementSymbol;
         residueIndex = this.atoms.residueIndex;
-        atomsDone = 0;
-        offset = 0;
-        bondsDone = 0;
 
-        atomsVector = new THREE.Vector3();
-        center = new THREE.Vector3();
-        rotationAxis = new THREE.Vector3();
-        up = new THREE.Vector3(0, 1, 0);
-        rotationAngle: number;
-
-        scaleMatrix = new THREE.Matrix4();
-        rotationMatrix = new THREE.Matrix4();
-        translationMatrix = new THREE.Matrix4();
-        atom1Vec = new THREE.Vector3();
-        atom2Vec = new THREE.Vector3();
-        finalMatrix: THREE.Matrix4;
-
+        scale = Vec3.zero();
+        translation = Vec3.zero();
+        
         pickColor = { r: 0.1, g: 0.1, b: 0.1 };
+        pickOffset = 0;
 
         atomMapBuilder = new Selection.VertexMapBuilder(this.atomIndices.length);
-        tempVector = new THREE.Vector3(0, 0, 0);
-
+        
         bs: BondsBuildState = <any>void 0;
 
         constructor(public model: Core.Structure.Molecule.Model, public atomIndices: number[], public params: Parameters) {
@@ -179,24 +171,13 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
         atomIndices = this.state.atomIndices;
         info = BallsAndSticksHelper.analyze(this.state.model, this.state.atomIndices, this.state.params);
 
-        //bondMapBuilder = new Selection.VertexMapBuilder(this.info.residueCount);
-
-        bondBufferSize = this.state.bondTemplateVertexBufferLength * this.info.stickCount;
-        bondVertices = new Float32Array(this.bondBufferSize);
-        bondNormals = new Float32Array(this.bondBufferSize);
-        bondColors = new Float32Array(this.bondBufferSize);
-        bondIndices = new Uint32Array(this.state.bondTemplateIndexBufferLength * this.info.stickCount);
-
+        bondVertexCount = this.state.bondTemplate.vertexCount * this.info.covalentStickCount + this.state.dashTemplate.vertexCount * this.info.metallicStickCount;
+        bondTriangleCount = this.state.bondTemplate.triangleCount * this.info.covalentStickCount + this.state.dashTemplate.triangleCount * this.info.metallicStickCount;
+        bondBuilder = GB.create(this.bondVertexCount, this.bondTriangleCount);
+        bondColors = new Float32Array(this.bondVertexCount * 3);
+        
         bondRadius = this.state.params.bondRadius;
-
-        //residueIndex = this.model.data.atoms.residueIndex;
-        //currentResidueIndex = this.residueIndex[this.info.bonds.atomAIndex[0]];
-        //bondMapVertexOffsetStart = 0;
-        //bondMapVertexOffsetEnd = 0;
-        covalentBondState = new BondModelState(
-            this.state.bondTemplate, this.state.bondTemplateVertexBuffer, this.state.bondTemplateNormalBuffer, this.state.bondTemplateIndexBuffer, this.state.bondTemplateVertexCount,
-            this.bondVertices, this.bondNormals, this.bondIndices);
-
+        bondState = new BondModelState(this.state.bondTemplate, this.state.dashTemplate, this.bondBuilder);
         bondCount = this.info.bonds.count;
 
         constructor(public state: BuildState) {
@@ -215,7 +196,6 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
 
         vertexStateBuffer: THREE.BufferAttribute;
 
-
         dispose() {
             this.atomsGeometry.dispose();
             this.bondsGeometry.dispose();
@@ -227,195 +207,123 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
 
             let r = state.atomRadius!(a);
 
-            state.scaleMatrix.makeScale(r, r, r);
-            state.tempVector.set(state.cX[a], state.cY[a], state.cZ[a]);
+            Vec3.set(state.scale, r, r, r);
+            Vec3.set(state.translation, state.cX[a], state.cY[a], state.cZ[a]);
 
-            state.translationMatrix.makeTranslation(state.tempVector.x, state.tempVector.y, state.tempVector.z).multiply(state.scaleMatrix);
-            state.atomTemplate.applyMatrix(state.translationMatrix);
-
-            for (let i = 0; i < state.atomTemplateVertexBufferLength; i++) {
-                state.offset = state.atomsDone * state.atomTemplateVertexBufferLength + i;
-                state.atomVertices[state.offset] = state.atomTemplateVertexBuffer[i];
-                state.atomNormals[state.offset] = state.atomTemplateNormalBuffer[i];
-            }
-
-            for (let i = 0; i < state.atomTemplateIndexBufferLength; i++) {
-                state.offset = state.atomTemplateIndexBufferLength * state.atomsDone + i;
-                state.atomTriangleIndices[state.offset] = state.atomTemplateIndexBuffer[i] + (state.atomTemplateVertexCount) * state.atomsDone;
-            }
+            const startVertexOffset = state.atomBuilder.vertexOffset;
+            GB.add(state.atomBuilder, state.atomTemplate, state.scale, state.translation, void 0);
 
             Selection.Picking.assignPickColor(a, state.pickColor);
 
-            for (let i = 0; i < state.atomTemplateVertexCount; i++) {
-                state.offset = state.atomsDone * state.atomTemplateVertexCount * 4 + 4 * i;
-                state.atomPickColors[state.offset] = state.pickColor.r;
-                state.atomPickColors[state.offset + 1] = state.pickColor.g;
-                state.atomPickColors[state.offset + 2] = state.pickColor.b;
+            for (let i = 0, _b = state.atomTemplate.vertexCount; i < _b; i++) {
+                state.atomPickColors[state.pickOffset++] = state.pickColor.r;
+                state.atomPickColors[state.pickOffset++] = state.pickColor.g;
+                state.atomPickColors[state.pickOffset++] = state.pickColor.b;
+                state.pickOffset++; // 4th component
             }
 
-            state.scaleMatrix.getInverse(state.translationMatrix);
-            state.atomTemplate.applyMatrix(state.scaleMatrix);
-
-            state.atomMapBuilder.addVertexRange(state.atomsDone * state.atomTemplateVertexCount, (state.atomsDone + 1) * state.atomTemplateVertexCount);
+            state.atomMapBuilder.addVertexRange(startVertexOffset, state.atomBuilder.vertexOffset);
             state.atomMapBuilder.endElement();
-
-            state.atomsDone++;
         }
         
         private static addBond(b: number, state: BuildState, bs: BondsBuildState) {
             let aI = bs.info.bonds.atomAIndex[b], bI = bs.info.bonds.atomBIndex[b], type = bs.info.bonds.type[b];
+            Vec3.set(bs.bondState.a, state.cX[aI], state.cY[aI], state.cZ[aI]);
+            Vec3.set(bs.bondState.b, state.cX[bI], state.cY[bI], state.cZ[bI]);
 
-            // if (bs.currentResidueIndex !== bs.residueIndex[aI]) {
-            //     bs.bondMapBuilder.addVertexRange(bs.bondMapVertexOffsetStart, bs.bondMapVertexOffsetEnd);
-            //     bs.bondMapVertexOffsetStart = bs.bondMapVertexOffsetEnd;
-            //     bs.bondMapBuilder.endElement();
-
-            //     bs.currentResidueIndex = bs.residueIndex[aI];
-            //     bs.bondMapBuilder.startElement(bs.currentResidueIndex);
-            // }
-
-            state.tempVector.set(state.cX[aI], state.cY[aI], state.cZ[aI]);
-            bs.covalentBondState.a.set(state.tempVector.x, state.tempVector.y, state.tempVector.z);
-
-            state.tempVector.set(state.cX[bI], state.cY[bI], state.cZ[bI]);
-            bs.covalentBondState.b.set(state.tempVector.x, state.tempVector.y, state.tempVector.z);
-
-            let r = +bs.bondRadius!, 
+            const r = +bs.bondRadius!, 
                 o = 2 * r / 3,
                 h = r / 2;
 
-            let bondState = bs.covalentBondState;
-            let order = 0;
+            const bondState = bs.bondState;
             switch (type) {
                 case BT.Unknown:
                 case BT.Single:
-                    order = 1;
                     bondState.radius = r!;
-                    bondState.offset.x = 0.0;
-                    bondState.offset.y = 0.0;
+                    bondState.offset[1] = 0.0;
+                    bondState.offset[2] = 0.0;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
                     break;
                 case BT.Double:
-                    order = 2;
                     bondState.radius = h;
-                    bondState.offset.x = o;
-                    bondState.offset.y = o;
+                    bondState.offset[1] = o;
+                    bondState.offset[2] = o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = -o;
-                    bondState.offset.y = -o;
+                    bondState.offset[1] = -o;
+                    bondState.offset[2] = -o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
                     break;
                 case BT.Triple:
-                    order = 3;
                     bondState.radius = h;
-                    bondState.offset.x = 0.0;
-                    bondState.offset.y = o;
+                    bondState.offset[1] = 0.0;
+                    bondState.offset[2] = o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = -Math.cos(Math.PI / 3) * o;
-                    bondState.offset.y = -Math.sin(Math.PI / 3) * o;
+                    bondState.offset[1] = -Math.cos(Math.PI / 3) * o;
+                    bondState.offset[2] = -Math.sin(Math.PI / 3) * o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = -bondState.offset.x;
-                    //bondState.offset.y = -0.05;
+                    bondState.offset[1] = -bondState.offset[0];
+                    //bondState.offset[1] = -0.05;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
                     break;
                 case BT.Aromatic:
-                    order = 4;
                     bondState.radius = h / 2;
-                    bondState.offset.x = o;
-                    bondState.offset.y = o;
+                    bondState.offset[1] = o;
+                    bondState.offset[2] = o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = -o;
-                    bondState.offset.y = -o;
+                    bondState.offset[1] = -o;
+                    bondState.offset[2] = -o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = -o;
-                    bondState.offset.y = o;
+                    bondState.offset[1] = -o;
+                    bondState.offset[2] = o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
-                    bondState.offset.x = o;
-                    bondState.offset.y = -o;
+                    bondState.offset[1] = o;
+                    bondState.offset[2] = -o;
                     BallsAndSticksGeometryBuilder.addBondPart(bondState);
                     break;
                 case BT.Metallic:
-                    order = 1;
-                    bondState.radius = r!;
-                    bondState.offset.x = 0.0;
-                    bondState.offset.y = 0.0;
-                    BallsAndSticksGeometryBuilder.addBondPart(bondState);
+                    bondState.radius = h;
+                    BallsAndSticksGeometryBuilder.addMetalBond(bondState);
                     break;
             }
-
-            //bs.bondMapVertexOffsetEnd += order * bs.state.bondTemplateVertexBufferLength;
         }
 
 
         private static addBondPart(state: BondModelState) {
-            state.atomsVector.subVectors(state.a, state.b);
+            const dir = Vec3.sub(state.dir, state.b, state.a);
+            const length = Vec3.length(state.dir);
+            const axis = Vec3.cross(state.rotationAxis, state.bondUpVector, dir);
+            const angle = Vec3.angle(state.bondUpVector, state.dir);
+            
 
-            let length = state.atomsVector.length();
+            Vec3.set(state.scale, length, state.radius, state.radius);
+            Mat4.fromRotation(state.rotation, angle, axis);
+            
+            state.offset[0] = 0;
+            Vec3.transformMat4(state.offset, state.offset, state.rotation);
+            Vec3.add(state.offset, state.offset, state.a);
 
-            state.center.addVectors(state.a, state.b).divideScalar(2);
-            state.rotationAxis.crossVectors(state.atomsVector, state.upVector).normalize();
-            let rotationAngle = state.atomsVector.angleTo(state.upVector);
-
-            state.scaleMatrix.makeScale(state.radius, length, state.radius);
-            state.offsetMatrix.makeTranslation(state.offset.x, state.offset.y, state.offset.z);
-            state.rotationMatrix.makeRotationAxis(state.rotationAxis, -rotationAngle);
-            state.translationMatrix.makeTranslation(state.center.x, state.center.y, state.center.z);
-
-            state.finalMatrix = state.translationMatrix.multiply(state.rotationMatrix).multiply(state.offsetMatrix).multiply(state.scaleMatrix);
-
-            state.template.applyMatrix(state.finalMatrix);
-
-            state.vertices.set(state.templateVB, state.templateVB.length * state.sticksDone);
-            state.normals.set(state.templateNB, state.templateVB.length * state.sticksDone);
-
-            let tIB = state.templateIB,
-                ib = state.indices,
-                offsetIB = state.templateIB.length * state.sticksDone,
-                offsetVB = state.templateVertexCount * state.sticksDone;
-            for (let i = 0; i < tIB.length; i++) {
-                ib[offsetIB++] = tIB[i] + offsetVB;
-            }
-
-            state.rotationMatrix.getInverse(state.finalMatrix);
-            state.template.applyMatrix(state.rotationMatrix);
-
-            state.sticksDone++;
+            GB.add(state.builder, state.bondTemplate, state.scale, state.offset, state.rotation);
         }
 
-        // private static addMetallicBondPart(state: BondModelState) {
-        //     state.atomsVector.subVectors(state.a, state.b);
+        private static addMetalBond(state: BondModelState) {
+            const dir = Vec3.sub(state.dir, state.b, state.a);
+            const length = Vec3.length(state.dir);
+            const axis = Vec3.cross(state.rotationAxis, state.dashUpVector, dir);
+            const angle = Vec3.angle(state.dashUpVector, state.dir);
 
-        //     let length = state.atomsVector.length();
-
-        //     state.center.addVectors(state.a, state.b).divideScalar(2);
-        //     state.rotationAxis.crossVectors(state.atomsVector, state.upVector).normalize();
-        //     let rotationAngle = state.atomsVector.angleTo(state.upVector);
-
-        //     state.scaleMatrix.makeScale(state.radius, length, state.radius);
-        //     state.offsetMatrix.makeTranslation(state.offset.x, state.offset.y, state.offset.z);
-        //     state.rotationMatrix.makeRotationAxis(state.rotationAxis, -rotationAngle);
-        //     state.translationMatrix.makeTranslation(state.center.x, state.center.y, state.center.z);
-
-        //     state.finalMatrix = state.translationMatrix.multiply(state.rotationMatrix).multiply(state.offsetMatrix).multiply(state.scaleMatrix);
-
-        //     state.template.applyMatrix(state.finalMatrix);
-
-        //     state.vertices.set(state.templateVB, state.templateVB.length * state.sticksDone);
-        //     state.normals.set(state.templateNB, state.templateVB.length * state.sticksDone);
-
-        //     let tIB = state.templateIB,
-        //         ib = state.indices,
-        //         offsetIB = state.templateIB.length * state.sticksDone,
-        //         offsetVB = state.templateVertexCount * state.sticksDone;
-        //     for (let i = 0; i < tIB.length; i++) {
-        //         ib[offsetIB++] = tIB[i] + offsetVB;
-        //     }
-
-        //     state.rotationMatrix.getInverse(state.finalMatrix);
-        //     state.template.applyMatrix(state.rotationMatrix);
-
-        //     state.sticksDone++;
-        // }
+            const scale = Vec3.set(state.scale, Constants.MetalDashSize, state.radius, state.radius);
+            const rotation = Mat4.fromRotation(state.rotation, angle, axis)!;
+            
+            const offset = state.offset;
+            Vec3.copy(offset, state.a);
+            Vec3.normalize(dir, dir);
+            Vec3.scale(dir, dir, 2 * Constants.MetalDashSize);
+            for (let t = 0; t < length; t += 2 * Constants.MetalDashSize) {
+                if (t + Constants.MetalDashSize > length) scale[0] = length - t;
+                GB.add(state.builder, state.cubeTemplate, scale, offset, rotation);
+                Vec3.add(offset, offset, dir);
+            }         
+        }
 
         private static getEmptyBondsGeometry() {
             let bondsGeometry = new THREE.BufferGeometry();
@@ -423,33 +331,32 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
             bondsGeometry.addAttribute('normal', new THREE.BufferAttribute(new Float32Array(0), 3));
             bondsGeometry.addAttribute('index', new THREE.BufferAttribute(new Uint32Array(0), 1));
             bondsGeometry.addAttribute('color', new THREE.BufferAttribute(new Float32Array(0), 3));
-            //let bondMapBuilder = new Selection.VertexMapBuilder(1);
-            return { bondsGeometry/*, bondVertexMap: bondMapBuilder.getMap()*/ };
+            return { bondsGeometry };
         }
 
         private static getBondsGeometry(state: BondsBuildState) {
             let bondsGeometry = new THREE.BufferGeometry();
-            bondsGeometry.addAttribute('position', new THREE.BufferAttribute(state.bondVertices, 3));
-            bondsGeometry.addAttribute('normal', new THREE.BufferAttribute(state.bondNormals, 3));
-            bondsGeometry.addAttribute('index', new THREE.BufferAttribute(state.bondIndices, 1));
+            bondsGeometry.addAttribute('position', new THREE.BufferAttribute(state.bondBuilder.vertices, 3));
+            bondsGeometry.addAttribute('normal', new THREE.BufferAttribute(state.bondBuilder.normals, 3));
+            bondsGeometry.addAttribute('index', new THREE.BufferAttribute(state.bondBuilder.indices, 1));
             bondsGeometry.addAttribute('color', new THREE.BufferAttribute(state.bondColors, 3));
-            return { bondsGeometry /*, bondVertexMap: state.bondMapBuilder.getMap()*/ };
+            return { bondsGeometry };
         }
 
         private static getAtomsGeometry(state: BuildState) {
             let atomsGeometry = new THREE.BufferGeometry();
-            atomsGeometry.addAttribute('position', new THREE.BufferAttribute(state.atomVertices, 3));
-            atomsGeometry.addAttribute('normal', new THREE.BufferAttribute(state.atomNormals, 3));
-            atomsGeometry.addAttribute('index', new THREE.BufferAttribute(state.atomTriangleIndices, 1));
+            atomsGeometry.addAttribute('position', new THREE.BufferAttribute(state.atomBuilder.vertices, 3));
+            atomsGeometry.addAttribute('normal', new THREE.BufferAttribute(state.atomBuilder.normals, 3));
+            atomsGeometry.addAttribute('index', new THREE.BufferAttribute(state.atomBuilder.indices, 1));
             atomsGeometry.addAttribute('color', new THREE.BufferAttribute(state.atomColors, 3));
 
-            let stateBuffer = new Float32Array(state.atomVertices.length);
+            let stateBuffer = new Float32Array(state.atomVertexCount);
             let vertexStateBuffer = new THREE.BufferAttribute(stateBuffer, 1);
             atomsGeometry.addAttribute('vState', vertexStateBuffer);
 
             let atomsPickGeometry = new THREE.BufferGeometry();
-            atomsPickGeometry.addAttribute('position', new THREE.BufferAttribute(state.atomVertices, 3));
-            atomsPickGeometry.addAttribute('index', new THREE.BufferAttribute(state.atomTriangleIndices, 1));
+            atomsPickGeometry.addAttribute('position', new THREE.BufferAttribute(state.atomBuilder.vertices, 3));
+            atomsPickGeometry.addAttribute('index', new THREE.BufferAttribute(state.atomBuilder.indices, 1));
             atomsPickGeometry.addAttribute('pColor', new THREE.BufferAttribute(state.atomPickColors, 4));
 
             return {
@@ -461,7 +368,7 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
         }
 
         private static async addAtoms(state: BuildState, ctx: Core.Computation.Context) {
-            const chunkSize = 1250;
+            const chunkSize = 2500;
             let started = Core.Utils.PerformanceMonitor.currentTime();
 
             for (let start = 0, _l = state.atomIndices.length; start < _l; start += chunkSize) {
@@ -477,7 +384,7 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
         }
 
         private static async addBondsChunks(state: BuildState, bs: BondsBuildState, ctx: Core.Computation.Context) {
-            const chunkSize = 1250;
+            const chunkSize = 2500;
             let started = Core.Utils.PerformanceMonitor.currentTime();
             for (let start = 0; start < bs.bondCount; start += chunkSize) {
                 for (let i = start, _b = Math.min(start + chunkSize, bs.bondCount); i < _b; i++) {
@@ -500,11 +407,7 @@ namespace LiteMol.Visualization.Molecule.BallsAndSticks {
 
             let bs = new BondsBuildState(state);
             state.bs = bs;
-            //bs.currentResidueIndex = bs.residueIndex[bs.info.bonds.atomAIndex[0]];
-            //bs.bondMapBuilder.startElement(bs.currentResidueIndex);
             await BallsAndSticksGeometryBuilder.addBondsChunks(state, bs, ctx);
-            //bs.bondMapBuilder.addVertexRange(bs.bondMapVertexOffsetStart, bs.bondMapVertexOffsetEnd);
-            //bs.bondMapBuilder.endElement();
         }
 
         static async build(model: Core.Structure.Molecule.Model,