+++ src/lib/Visualization/Utils/GeometryBuilder.ts
@@ -8,8 +8,7 @@ namespace LiteMol.Visualization.Utils {
     export interface GeometryBuilder {
         vertices: Float32Array,
         normals: Float32Array,
-        indices: Uint32Array,
-        annotation: Int32Array | undefined,
+        indices: Uint32Array
 
         vertexOffset: number,
         indexOffset: number
@@ -20,26 +19,25 @@ namespace LiteMol.Visualization.Utils {
         import Vec3 = Geom.LinearAlgebra.Vector3
         import Mat4 = Geom.LinearAlgebra.Matrix4
 
-        export function create(vertexCount: number, triangleCount: number, includeAnnotation: boolean): GeometryBuilder {            
+        export function create(vertexCount: number, triangleCount: number): GeometryBuilder {            
             return {
                 vertices: new Float32Array(vertexCount * 3),
                 normals: new Float32Array(vertexCount * 3),
                 indices: new Uint32Array(triangleCount * 3),
-                annotation: includeAnnotation ? new Int32Array(vertexCount * 3) : void 0,
                 vertexOffset: 0,
                 indexOffset: 0
             }
         }
 
         const tempVector = Vec3.zero();
-        function copyTransformed(builder: GeometryBuilder, surface: Geom.Surface, vertTransform: Mat4, normalTransform: Mat4, constAnnotation: number | undefined) {
-            const { vertices, normals, indices, vertexOffset, indexOffset, annotation } = builder;
-            const { vertices: vs, normals: ns, triangleIndices: ts, annotation: ans, vertexCount } = surface;
+        function copyTransformed(builder: GeometryBuilder, surface: Geom.Surface, vertTransform: Mat4, normalTransform: Mat4) {
+            const { vertices, normals, indices, vertexOffset, indexOffset } = builder;
+            const { vertices: vs, normals: ns, triangleIndices: ts, vertexCount, triangleCount } = surface;
             const v = tempVector;
 
-            let vOffset = vertexOffset;
-            let nOffset = vertexOffset;
-            let iOffset = indexOffset;
+            let vOffset = 3 * vertexOffset;
+            let nOffset = 3 * vertexOffset;
+            let iOffset = 3 * indexOffset;
 
             for (let i = 0, _b = vs.length; i < _b; i += 3) {
                 v[0] = vs[i], v[1] = vs[i + 1], v[2] = vs[i + 2];
@@ -57,31 +55,18 @@ namespace LiteMol.Visualization.Utils {
             }
 
             for (let i = 0, _b = ts.length; i < _b; i++) {
-                indices[iOffset++] = ts[i];
+                indices[iOffset++] = vertexOffset + ts[i];
             }
 
-            if (ans && annotation) {
-                let aOffset = vertexOffset;
-                for (let i = 0, _b = ts.length; i < _b; i++) {
-                    annotation[aOffset++] = ans[i];
-                }
-            } else if (constAnnotation !== void 0 && annotation) {
-                let aOffset = vertexOffset;
-                for (let i = 0, _b = vertexCount; i < _b; i ++) {
-                    annotation[aOffset++] = constAnnotation;
-                }
-            }
-
-            builder.vertexOffset = vOffset;
-            builder.indexOffset = iOffset;
+            builder.vertexOffset += vertexCount;
+            builder.indexOffset += triangleCount;
         }
 
         const scaleTransform = Mat4.zero(), translateTransform = Mat4.zero(), rotateTransform = Mat4.zero(), vTransform = Mat4.zero(), nTransform = Mat4.zero();
         const defaulScale = [1, 1, 1], defaultTranslation = [0, 0, 0];
         export function add(
             builder: GeometryBuilder, surface: Geom.Surface, 
-            scale: number[] | undefined, translation: number[] | undefined, rotation: Mat4 | undefined, 
-            annotation: number | undefined) {
+            scale: number[] | undefined, translation: number[] | undefined, rotation: Mat4 | undefined) {
 
             Mat4.fromScaling(scaleTransform, scale || defaulScale);
             Mat4.fromTranslation(translateTransform, translation || defaultTranslation);
@@ -91,7 +76,7 @@ namespace LiteMol.Visualization.Utils {
             Mat4.mul3(vTransform, translateTransform, rotateTransform, scaleTransform);
             Mat4.mul(nTransform, rotateTransform, scaleTransform);
 
-            copyTransformed(builder, surface, vTransform, nTransform, annotation);
+            copyTransformed(builder, surface, vTransform, nTransform);
         }
     }
 }