+++ src/lib/Visualization/CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿# 1.7.1
 * MouseWheel event is now triggered only if the user moves the mouse over the area.
+* Refactored balls and sticks model to include metallic bonds.
 
 # 1.7.0
 * Added support for labels.